refactor: change package name - error -> exception
