+++ src/main/java/kakao/exception/CustomExceptionHandler.java
@@ -1,4 +1,4 @@
-package kakao.error;
+package kakao.exception;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
