+++ src/main/java/kakao/exception/CustomRuntimeException.java
@@ -1,4 +1,4 @@
-package kakao.error;
+package kakao.exception;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
